+++ example/lib/examples/controls.dart
@@ -13,9 +13,30 @@ class ControlExample extends StatefulWidget {
 class _ControlExampleState extends State<ControlExample> {
   @override
   Widget build(BuildContext context) {
-    return SkadiScaffold(
-      allowPop: true,
-      body: ExampleScaffold(
+    return WillPopScope(
+      onWillPop: () async {
+        bool isLoading = LoadingOverlayProvider.instance.isLoading;
+        if (isLoading) {
+          LoadingOverlayProvider.switchPosition(LoadingOverlayPosition.below);
+          bool? result = await showDialog(
+            context: context,
+            barrierDismissible: false,
+            builder: (context) => const SkadiConfirmationDialog.danger(
+              content: Text("Are you sure you want to cancel this operation?"),
+              confirmText: "Yes",
+              cancelText: "Discard",
+            ),
+          );
+          if (result == true) {
+            ///Cancel your operation
+            LoadingOverlayProvider.toggle(false);
+          }
+          LoadingOverlayProvider.switchPosition(LoadingOverlayPosition.above);
+          return false;
+        }
+        return true;
+      },
+      child: ExampleScaffold(
         title: "Control",
         children: [
           Section(
@@ -51,7 +72,7 @@ class _ControlExampleState extends State<ControlExample> {
               ElevatedButton(
                 onPressed: () async {
                   LoadingOverlayProvider.toggle();
-                  await SkadiUtils.wait(21600);
+                  await SkadiUtils.wait(10000);
                   LoadingOverlayProvider.toggle(false);
                 },
                 child: const Text("Click"),