+++ lib/src/widgets/loading_overlay.dart
@@ -1,9 +1,15 @@
 import 'package:flutter/material.dart';
 
-import '../../skadi.dart';
+enum LoadingOverlayPosition {
+  above,
+  below,
+}
 
-class LoadingOverlayProvider {
-  final ValueNotifier<bool> _loadingNotifier = ValueNotifier(false);
+class LoadingOverlayProvider extends ChangeNotifier {
+  bool _loadingNotifier = false;
+  LoadingOverlayPosition _overlayPosition = LoadingOverlayPosition.above;
+
+  //
   LoadingOverlayProvider._();
   static LoadingOverlayProvider instance = LoadingOverlayProvider._();
 
@@ -15,15 +21,29 @@ class LoadingOverlayProvider {
     );
   }
 
-  bool get isLoading => _loadingNotifier.value;
+  bool get isLoading => _loadingNotifier;
+  LoadingOverlayPosition get position => _overlayPosition;
 
+  ///Toggle the loading overlay
   static void toggle([bool? value]) {
     LoadingOverlayProvider.instance._toggleLoading(value);
   }
 
+  ///Bring loading overlay below our current widget
+  ///useful when you want to show dialog while loading
+  static void switchPosition(LoadingOverlayPosition position) {
+    LoadingOverlayProvider.instance._switchPosition(position);
+  }
+
   void _toggleLoading([bool? value]) {
-    final bool isLoading = value ?? !_loadingNotifier.value;
-    _loadingNotifier.value = isLoading;
+    final bool isLoading = value ?? !_loadingNotifier;
+    _loadingNotifier = isLoading;
+    notifyListeners();
+  }
+
+  void _switchPosition(LoadingOverlayPosition position) {
+    _overlayPosition = position;
+    notifyListeners();
   }
 }
 
@@ -46,28 +66,33 @@ class _LoadingOverlayBuilder extends StatelessWidget {
     final color = Theme.of(context).brightness == Brightness.dark
         ? Colors.grey.withOpacity(0.2)
         : Colors.black26;
-    return Stack(
-      children: [
-        child,
-        ValueListenableBuilder<bool>(
-          valueListenable: LoadingOverlayProvider.instance._loadingNotifier,
-          child: loadingWidget ??
-              Container(
-                color: color,
-                child: const Center(
-                  child: CircularProgressIndicator(
-                    backgroundColor: Colors.white,
-                  ),
-                ),
-              ),
-          builder: (context, isLoading, child) {
-            if (isLoading) {
-              return child!;
-            }
-            return emptySizedBox;
-          },
-        ),
-      ],
+    final instance = LoadingOverlayProvider.instance;
+    final loadingContent = loadingWidget ??
+        Container(
+          color: color,
+          child: const Center(
+            child: CircularProgressIndicator(
+              backgroundColor: Colors.white,
+            ),
+          ),
+        );
+    return AnimatedBuilder(
+      animation: instance,
+      builder: (context, child) {
+        final bool isLoading = instance.isLoading;
+        return Stack(
+          children: [
+            if (instance._overlayPosition == LoadingOverlayPosition.above) ...[
+              child!,
+              if (isLoading) loadingContent,
+            ] else ...[
+              if (isLoading) loadingContent,
+              child!,
+            ],
+          ],
+        );
+      },
+      child: child,
     );
   }
 }