+++ lib/src/widgets/skadi_scaffold.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:skadi/skadi.dart';
 
 class SkadiScaffold extends StatelessWidget {
-  final AppBar? appBar;
+  final PreferredSizeWidget? appBar;
   final Widget? body;
   final bool allowPop;
   final Scaffold? scaffold;